+++ react-components/src/pages/CardsPage/CardsPage.module.scss
@@ -1,8 +1,12 @@
 .cardWrapper {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-    gap: 20px;
-    align-items: stretch;
-    width: 100%;
-  
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  gap: 20px;
+  align-items: stretch;
+  width: 100%;
+}
+
+.notFound {
+  font-weight: 900;
+  font-size: 20px;
 }
